Refresher: split classes in different modules for better handling of tornado import
